IC: déplacement du code inclusion connect dans dora.oidc
