+++ requirements/dev.txt
@@ -11,3 +11,4 @@ ruff==0.1.4
 pytest==7.4.3
 pytest-django==4.6.0
 sqlfluff==2.3.5
+requests-mock==1.11.0
